refactor: move and merge some files (#1419)
