+++ frontend/src/components/Navbar/Logo.tsx
@@ -4,7 +4,7 @@ import { useTheme } from 'styled-components';
 import logo from '../../images/brand/bluebook.svg';
 import wordmark_outlines from '../../images/brand/wordmark_outlines.svg';
 import wordmark_outlines_dark from '../../images/brand/ct_white.svg';
-import common_styles from '../../styles/common.module.css';
+import styles from './Logo.module.css';
 
 type Props = {
   /** Should we show the icon */
@@ -21,13 +21,13 @@ function Logo({ icon = true, wordmark = true }: Props) {
 
   return (
     <span
-      className={common_styles.coursetable_logo}
+      className={styles.coursetable_logo}
       style={{
         display: 'block',
       }}
     >
       {icon && (
-        <img src={logo} alt="" className={common_styles.coursetable_logo_img} />
+        <img src={logo} alt="" className={styles.coursetable_logo_img} />
       )}{' '}
       {wordmark && (
         <img
@@ -35,7 +35,7 @@ function Logo({ icon = true, wordmark = true }: Props) {
             theme.theme === 'dark' ? wordmark_outlines_dark : wordmark_outlines
           }
           alt="CourseTable"
-          className={common_styles.coursetable_logo_wordmark}
+          className={styles.coursetable_logo_wordmark}
         />
       )}
     </span>
