+++ frontend/src/components/Navbar/Navbar.tsx
@@ -7,7 +7,7 @@ import styled from 'styled-components';
 import Logo from './Logo';
 import DarkModeButton from './DarkModeButton';
 import MeDropdown from './MeDropdown';
-import { useWindowDimensions } from '../Providers/WindowDimensionsProvider';
+import { useWindowDimensions } from '../../contexts/windowDimensionsContext';
 import {
   breakpoints,
   logout,
