--- frontend/src/components/UseDarkMode.tsx
@@ -1,23 +1,0 @@
-import { useEffect, useState } from 'react';
-
-type Theme = 'light' | 'dark';
-
-export const useDarkMode = (): [Theme, () => void] => {
-  const [theme, setTheme] = useState<Theme>('light');
-
-  const setMode = (mode: Theme) => {
-    window.localStorage.setItem('theme', mode);
-    setTheme(mode);
-  };
-
-  const themeToggler = () => {
-    if (theme === 'light') setMode('dark');
-    else setMode('light');
-  };
-
-  useEffect(() => {
-    const localTheme = window.localStorage.getItem('theme') as Theme;
-    if (localTheme) setTheme(localTheme);
-  }, []);
-  return [theme, themeToggler];
-};
