--- frontend/src/env.d.ts
@@ -1,9 +1,0 @@
-interface ImportMetaEnv extends Readonly<Record<string, string>> {
-  readonly VITE_API_ENDPOINT: string;
-  readonly VITE_CANNY_ID: string;
-  readonly VITE_SENTRY_RELEASE: string;
-}
-
-interface ImportMeta {
-  readonly env: ImportMetaEnv;
-}
