+++ frontend/src/utilities/browserStorage.ts
