+++ frontend/src/vite-env.d.ts
@@ -1,1 +1,11 @@
 /// <reference types="vite/client" />
+
+interface ImportMetaEnv extends Readonly<Record<string, string>> {
+  readonly VITE_API_ENDPOINT: string;
+  readonly VITE_CANNY_ID: string;
+  readonly VITE_SENTRY_RELEASE: string;
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv;
+}
